Split up system control from interrupt
- Update call to `is_a_test()` calls in code
- Add a `enable_floating_point_unit()` to system_control class
